chill_requirements.txt
backports.tarfile==1.2.0
coverage==7.6.7
discord-webhook==1.3.1
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
duckduckgo-search==6.3.7
gunicorn==23.0.0
importlib-metadata==8.0.0
jaraco.collections==5.1.0
langchain-community==0.3.10
langchain-groq==0.2.1
langchain-ollama==0.2.1
langchain-postgres==0.0.12
langgraph==0.2.56
markdown==3.7
maturin==1.7.8
pillow==11.0.0
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pygraphviz==1.14
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-memcached==1.62
rust-lib==0.1.0
tiktoken==0.8.0
tomli==2.0.1
